+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGES
 **3.0.2**
 * Support day field for TimePicker
+* Fixed error when typing in hh:mm format
 
 **3.0.1**
 * Support textarea for TimePicker